Refactor matrix loading into matrix infrastructure layer
