+++ src/monocycle_nash/matrix/__init__.py
@@ -3,6 +3,7 @@ from .general import GeneralPayoffMatrix
 from .monocycle import MonocyclePayoffMatrix
 from .builder import PayoffMatrixBuilder
 from .random import RandomMatrixAcceptanceCondition, generate_random_skew_symmetric_matrix
+from .infra import MatrixFileInfrastructure, build_characters, build_matrix_from_input, has_matrix_input, validate_matrix_input
 from .approximation import (
     PayoffMatrixApproximation,
     MonocycleToGeneralApproximation,
@@ -29,4 +30,9 @@ __all__ = [
     "MaxElementDifferenceDistance",
     "EquilibriumUStrategyDifferenceDistance",
     "ApproximationQualityEvaluator",
+    "MatrixFileInfrastructure",
+    "build_matrix_from_input",
+    "build_characters",
+    "has_matrix_input",
+    "validate_matrix_input",
 ]
